requirements.txt
pandas
numpy
matplotlib
seaborn
tiktoken
tqdm
openai
anthropic
google-cloud-aiplatform
python-dotenv